requirements.txt
pytest==8.3.5
mypy==1.15.0
black==25.1.0
pytest-cov==6.1.1
python-dotenv==1.2.1
openai==2.15.0
